backend/requirements.txt
Flask==3.0.3                # Framework utama untuk web
pymongo==4.7.1             # Untuk koneksi MongoDB
flask-cors==4.0.1         # Untuk mengaktifkan CORS (Cross-Origin Resource Sharing)
pandas==2.2.2              # Untuk manipulasi data dan ekspor data ke CSV
reportlab==4.2.0          # Untuk ekspor PDF
xlswriter==3.2.0
python-dotenv==1.0.1